+++ com-tech-api/src/main/java/com/tech/api/storage/model/Customer.java
@@ -35,6 +35,6 @@ public class Customer extends Auditable<String> {
     private List<CustomerAddress> customerAddresses;
 
     private Integer loyaltyLevel = 0;   //0: bronze, 1: silver, 2: gold, 3: platinum, 4: diamond, 5: black diamond
-    private Double loyaltyPoint = 0d;
-    private Double point = 0d;
+    private Integer loyaltyPoint = 0;
+    private Integer point = 0;
 }
